sys: add secure memory functions
the use case for these functions is secure code like
SRTP and TLS.

memcmp() should not be used when comparing secret data
such as shared keys, because the content may affect the timing.
instead the compare function should always take the same
to execute, independently of the content.

move it to mem module