rtmp: update timestamp after complete packet
if the RTMP packet received is not complete, the dechunker
will decode the RTMP header and then return (wait for more data).

currently it erroneously also updates the timestamp from the delta,
even if the packet is incomplete. this means that the timestamp
will be updated twice, which is not correct.

a suggested fix is to move the timestamp calculation to after the
complete packet has been read.