Honour interface preference
An interface or address preference can be set by clients.  However, this
is not always honoured - instead leaving it to the OS to choose the
interface on which to send the packets.  This is usually The Right Thing
to do as the OS is in the best place to choose the interface that will
most likely get the packet to the destination via the ip address / route
metric mechanisms.

However, in certain use cases where a clear preference has been
identified by the client, libre should be doing its best to use that as
the outgoing interface.  This can be achieved in two ways (that I know
of).

1) Set the socket option SO_BINDTODEVICE which requires net admin
privileges
or
2) Call bind() on the socket with the desired address which will set the
src address in the packets correctly, and it is then up to the OS for
which physical interface the packets are sent on, but with the src
address set, it is much easier to configure the OS to always send
those packets to a particular device.