http/client: add set functions for client certificate and private key
Adds functions that sets the client certificate and private key for HTTPS. The
certificate and key can be specified as:
- file with http_client_set_cert(), http_client_set_key() or
- directly as PEM string with http_client_set_certpem(),
  http_client_set_keypem().
The function http_client_set_cert() and http_client_set_certpem() accept PEM
file/data where certificate and private key are combined or only certificate
is contained. If the key is not contained then also http_client_set_key() or
http_client_set_keypem() has to be used.